GRAD2-2110: refactored to use processor to avoid transaction timeout for distribution job status update.
GRAD2-2110: refactored to use processor to avoid transaction timeout for distribution job status update.
